+++ gym_space_engineers/__init__.py
@@ -56,7 +56,7 @@ register(
         "add_end_effector_velocity": True,
         "task": "forward",
         "randomize_task": True,
-        "randomize_interval": 50, # change task every 5s
+        "randomize_interval": 100, # change task every 10s
     },
 )