+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Release 0.0.6
 - Added intra episode task randomization
+- Scaled reward for the turning task
 
 # Release 0.0.5
 - New convention for end effector z